🚚 chore: rename contacts schema to links
