+++ src/components/layout/LinkNav.astro
@@ -1,5 +1,6 @@
 ---
-import processProfileLinks from '@/schemas/contacts'
+import { processProfileLinks } from '@/schemas'
+
 import { profileLinks } from '@/site-config'
 
 interface Props {
