+++ src/content.config.ts
@@ -1,8 +1,7 @@
 import { defineCollection } from 'astro:content'
 import { glob } from 'astro/loaders'
 
-import { blogSchema } from './schemas/blog'
-import { projectsSchema } from './schemas/projects'
+import { blogSchema, projectsSchema } from './schemas'
 
 // Blog collection
 const blog = defineCollection({
